requirements.txt
accelerate==0.20.3
adapter-transformers==3.2.1
bitsandbytes==0.39.0
ctranslate2==3.15.1
datasets==2.12.0
fastapi==0.95.2
gradio==3.31.0
loralib==0.1.1
protobuf==3.20.0
scipy==1.10.1
sentencepiece==0.1.99
torch==2.0.1 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.2 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118
git+https://github.com/huggingface/peft.git
